add repository config for spring-ai bom

+++ src/spring-petclinic-chat-agent/src/main/java/org/springframework/samples/petclinic/agent/chat/ChatOptionsProperties.java
@@ -13,6 +13,6 @@ class ChatOptionsProperties {
 
     String deploymentName;
 
-    Float temperature;
+    Double temperature;
 
 }
